+++ CHANGELOG.md
@@ -9,8 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Automatic title change according to ROM name
-* Support for emulator sections
+*
 
 ### Changed
 
@@ -19,6 +18,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+
+## [0.8.4] - 2023-04-20
+
+### Added
+
+* Automatic title change according to ROM name
+* Support for emulator sections
 
 ## [0.8.3] - 2023-04-11